requirements.txt
flask==2.3.2
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
tensorflow==2.13.0
joblib==1.3.2